Order, Payment Checkout completed

+++ Frontends/ECommerce.Web/Models/Cart/CartItemViewModel.cs
@@ -6,7 +6,7 @@ public class CartItemViewModel
     public string? ProductId { get; set; }
     public string? Name { get; set; }
     public decimal Price { get; set; }
-    public int Quantity { get; set; }
+    public byte Quantity { get; set; }
 
     private decimal? DiscountPrice;
     public decimal GetCurrentPrice
